Pre v0.1.0 cleanup (#88)
* remove _eul in eurec4a circle domain name
* remove unused files
* minor refactor for distance calcs, move into module with geometry and remove duplicate module
