+++ input_examples/domains/eurec4a_circle.yaml
