--- lagtraj/utils/dephy.py
@@ -1,5 +1,0 @@
-# Utilities related to DEPHY data
-# TODO
-# - Add list of variables
-
-dephy_variables = {}
