--- lagtraj/utils/settings.py
@@ -1,1 +1,0 @@
-from yaml import *
